GooglePlayService implemented
